+++ src/Primitives/MeshGenerators/Arc.cpp
@@ -1,7 +1,7 @@
 #include "../Mesh.h"
 #include <Essentials/Tools.h>
 
-Mesh* Mesh::generateArc(const vec2& dimensions, const double& angle, const unsigned int& resolution)
+Mesh* Mesh::generateArc(const vec2& dimensions, const float& angle, const unsigned int& resolution)
 {
     std::string name = "INTERNAL_ARC_" + std::to_string(angle) + "_" + std::to_string(resolution);
     if(Tools::mapHasKey(mLoadedMeshes, name))
@@ -9,14 +9,14 @@ Mesh* Mesh::generateArc(const vec2& dimensions, const double& angle, const unsig
 
     Mesh* mesh = new Mesh(name);
 
-    double stepsize = angle / (double)resolution;
-    for(double i = 0; i < angle; i += stepsize)
+    float stepsize = angle / (float)resolution;
+    for(float i = 0; i < angle; i += stepsize)
     {
         Vertex vert;
         vert.position = vec3(
-            cos(Gum::Maths::toRadians(i)) * dimensions.x,
+            cosf(Gum::Maths::toRadians(i)) * dimensions.x,
             0.0f,
-            sin(Gum::Maths::toRadians(i)) * dimensions.y
+            sinf(Gum::Maths::toRadians(i)) * dimensions.y
         );
         mesh->addVertex(vert);
     }
